+++ packages/react-components/src/stories/DfxIcon.tsx
@@ -1,5 +1,5 @@
 import { IconContext, IconType } from 'react-icons';
-import { FaCircle, FaRegCircle } from 'react-icons/fa';
+import { FaCircle, FaRegCircle, FaUserCog, FaFileAlt, FaTelegramPlane } from 'react-icons/fa';
 import {
   MdContentCopy,
   MdArrowBackIos,
@@ -27,6 +27,8 @@ import {
   MdMoreVert,
   MdOpenInNew,
   MdSearch,
+  MdHome,
+  MdMenu,
 } from 'react-icons/md';
 import { VscArrowSwap } from 'react-icons/vsc';
 import { ReactElement, useContext } from 'react';
@@ -38,6 +40,7 @@ interface DfxIconProps {
 }
 
 export enum IconVariant {
+  ACCOUNT = 'ACCOUNT',
   ARROW_RIGHT = 'ARROW_RIGHT',
   ARROW_LEFT = 'ARROW_LEFT',
   BACK = 'BACK',
@@ -53,10 +56,13 @@ export enum IconVariant {
   EDIT = 'EDIT',
   EXPAND_LESS = 'EXPAND_LESS',
   EXPAND_MORE = 'EXPAND_MORE',
+  FILE = 'FILE',
   FORWARD = 'FORWARD',
   HELP = 'HELP',
+  HOME = 'HOME',
   INFO = 'INFO',
   INFO_OUTLINE = 'INFO_OUTLINE',
+  MENU = 'MENU',
   PROCESS_DONE = 'PROCESS_DONE',
   SEPA_INSTANT = 'SEPA_INSTANT',
   SETTINGS = 'SETTINGS',
@@ -66,6 +72,7 @@ export enum IconVariant {
   CHECKBOX_EMPTY = 'CHECKBOX_EMPTY',
   CHECKBOX_CHECKED = 'CHECKBOX_CHECKED',
   SIGNATURE_POPUP = 'SIGNATURE_POPUP',
+  TELEGRAM = 'TELEGRAM',
   THREE_DOTS_VERT = 'THREE_DOTS_VERT',
   OPEN_IN_NEW = 'OPEN_IN_NEW',
   METAMASK_LOGO = 'METAMASK_LOGO',
@@ -78,6 +85,8 @@ export enum IconVariant {
 }
 
 export const VARIANT_MAPS: Record<IconVariant, ReactElement<IconType>> = {
+  [IconVariant.MENU]: <MdMenu />,
+  [IconVariant.ACCOUNT]: <FaUserCog />,
   [IconVariant.COPY]: <MdContentCopy />,
   [IconVariant.BACK]: <MdArrowBackIos />,
   [IconVariant.FORWARD]: <MdArrowForwardIos />,
@@ -85,6 +94,7 @@ export const VARIANT_MAPS: Record<IconVariant, ReactElement<IconType>> = {
   [IconVariant.UNFOLD_MORE]: <MdUnfoldMore />,
   [IconVariant.EXPAND_LESS]: <MdExpandLess />,
   [IconVariant.EXPAND_MORE]: <MdExpandMore />,
+  [IconVariant.FILE]: <FaFileAlt />,
   [IconVariant.CHECK]: <MdCheck />,
   [IconVariant.CHECKBOX_EMPTY]: <MdOutlineCheckBoxOutlineBlank />,
   [IconVariant.CHECKBOX_CHECKED]: <MdOutlineCheckBox />,
@@ -99,6 +109,7 @@ export const VARIANT_MAPS: Record<IconVariant, ReactElement<IconType>> = {
   [IconVariant.SETTINGS]: <MdSettings />,
   [IconVariant.WALLET]: <DfxWalletIcon />,
   [IconVariant.BANK]: <DfxBankIcon />,
+  [IconVariant.HOME]: <MdHome />,
   [IconVariant.SEPA_INSTANT]: <DfxSepaInstantAvailable />,
   [IconVariant.PROCESS_DONE]: <DfxProcessDoneIcon />,
   [IconVariant.EDIT]: <MdEdit />,
@@ -115,6 +126,7 @@ export const VARIANT_MAPS: Record<IconVariant, ReactElement<IconType>> = {
   [IconVariant.CIRCLE]: <FaCircle />,
   [IconVariant.CIRCLE_OUTLINE]: <FaRegCircle />,
   [IconVariant.USER_DATA]: <DfxUserData />,
+  [IconVariant.TELEGRAM]: <FaTelegramPlane />,
 };
 
 export enum IconSize {