+++ packages/react/src/contexts/dfx.context.tsx
@@ -5,6 +5,7 @@ import { BankAccountContextProvider } from './bank-account.context';
 import { SessionContextProvider, SessionContextProviderProps } from './session.context';
 import { UserContextProvider } from './user.context';
 import { FiatContextProvider } from './fiat.context';
+import { LanguageContextProvider } from './language.context';
 
 type DfxContextProviderProps = SessionContextProviderProps & PropsWithChildren;
 
@@ -14,9 +15,11 @@ export function DfxContextProvider(props: DfxContextProviderProps): JSX.Element 
       <SessionContextProvider api={props.api} data={props.data}>
         <UserContextProvider>
           <FiatContextProvider>
-            <AssetContextProvider>
-              <BankAccountContextProvider>{props.children}</BankAccountContextProvider>
-            </AssetContextProvider>
+            <LanguageContextProvider>
+              <AssetContextProvider>
+                <BankAccountContextProvider>{props.children}</BankAccountContextProvider>
+              </AssetContextProvider>
+            </LanguageContextProvider>
           </FiatContextProvider>
         </UserContextProvider>
       </SessionContextProvider>