+++ packages/react/src/definitions/buy.ts
@@ -5,25 +5,30 @@ export const BuyUrl = { receive: 'buy/paymentInfos' };
 
 export interface Buy {
   name: string;
-  country: string;
   street: string;
-  city: string;
-  zip: string;
   number: string;
-  fee: number;
+  zip: string;
+  city: string;
+  country: string;
   iban: string;
   bic: string;
+  sepaInstant: boolean;
+  routeId: number;
   remittanceInfo: string;
+  fee: number;
   minFee: number;
   minVolume: number;
   minFeeTarget: number;
   minVolumeTarget: number;
+  amount: number;
+  currency: Fiat;
   estimatedAmount: number;
-  sepaInstant: boolean;
+  asset: Asset;
 }
 
 export interface BuyPaymentInfo {
   currency: Fiat;
-  amount: number;
+  amount?: number;
   asset: Asset;
+  targetAmount?: number;
 }