+++ packages/react/src/hooks/asset.hook.ts
@@ -1,8 +1,10 @@
+import { useMemo } from 'react';
 import { Asset, AssetUrl } from '../definitions/asset';
 import { useApi } from './api.hook';
 
 export interface AssetInterface {
   getAssets: () => Promise<Asset[]>;
+  getAsset: (assets?: Asset[], identifier?: string) => Asset | undefined;
 }
 
 export function useAsset(): AssetInterface {
@@ -12,5 +14,15 @@ export function useAsset(): AssetInterface {
     return call<Asset[]>({ url: AssetUrl.get, method: 'GET' });
   }
 
-  return { getAssets };
+  function getAsset(assets: Asset[] = [], identifier?: string): Asset | undefined {
+    if (!identifier) return undefined;
+
+    return (
+      assets.find((a) => a.id === +identifier) ??
+      assets.find((a) => a.uniqueName.toLowerCase() === identifier.toLowerCase()) ??
+      assets.find((a) => a.name.toLowerCase() === identifier.toLowerCase())
+    );
+  }
+
+  return useMemo(() => ({ getAssets, getAsset }), [call]);
 }