+++ packages/react/src/hooks/auth.hook.ts
@@ -1,10 +1,18 @@
+import { useMemo } from 'react';
 import { AuthUrl, SignIn, SignMessage } from '../definitions/auth';
 import { useApi } from './api.hook';
 
 export interface AuthInterface {
   getSignMessage: (address: string) => Promise<string>;
   signIn: (address: string, signature: string) => Promise<SignIn>;
-  signUp: (address: string, signature: string, walletId?: number) => Promise<SignIn>;
+  signUp: (address: string, signature: string, wallet?: string) => Promise<SignIn>;
+}
+
+interface SignUpParams {
+  address: string;
+  signature: string;
+  walletId?: number;
+  wallet?: string;
 }
 
 export function useAuth(): AuthInterface {
@@ -20,9 +28,20 @@ export function useAuth(): AuthInterface {
     return call({ url: AuthUrl.signIn, method: 'POST', data: { address, signature } });
   }
 
-  async function signUp(address: string, signature: string, walletId?: number): Promise<SignIn> {
-    return call({ url: AuthUrl.signUp, method: 'POST', data: { address, signature, walletId } });
+  async function signUp(address: string, signature: string, wallet?: string): Promise<SignIn> {
+    const data: SignUpParams = { address, signature };
+
+    if (wallet) {
+      const walletId = parseInt(wallet);
+      if (!isNaN(walletId)) {
+        data.walletId = walletId;
+      } else {
+        data.wallet = wallet;
+      }
+    }
+
+    return call({ url: AuthUrl.signUp, method: 'POST', data });
   }
 
-  return { getSignMessage, signIn, signUp };
+  return useMemo(() => ({ getSignMessage, signIn, signUp }), [call]);
 }