+++ packages/react/src/hooks/bank-account.hook.ts
@@ -1,3 +1,4 @@
+import { useMemo } from 'react';
 import { BankAccount, BankAccountUrl } from '../definitions/bank-account';
 import { Fiat } from '../definitions/fiat';
 import { useApi } from './api.hook';
@@ -15,6 +16,7 @@ export interface UpdateBankAccount {
 
 export interface BankAccountInterface {
   getAccounts: () => Promise<BankAccount[]>;
+  getAccount: (accounts?: BankAccount[], identifier?: string) => BankAccount | undefined;
   createAccount: (newAccount: CreateBankAccount) => Promise<BankAccount>;
   updateAccount: (id: number, changedAccount: UpdateBankAccount) => Promise<BankAccount>;
 }
@@ -26,6 +28,16 @@ export function useBankAccount(): BankAccountInterface {
     return call<BankAccount[]>({ url: BankAccountUrl.get, method: 'GET' });
   }
 
+  function getAccount(accounts: BankAccount[] = [], identifier?: string): BankAccount | undefined {
+    if (!identifier) return undefined;
+
+    return (
+      accounts.find((b) => b.id === +identifier) ??
+      accounts.find((b) => b.iban.toLowerCase() === identifier.toLowerCase()) ??
+      accounts.find((b) => b.label?.toLowerCase() === identifier.toLowerCase())
+    );
+  }
+
   async function createAccount(newAccount: CreateBankAccount): Promise<BankAccount> {
     return call<BankAccount>({ url: BankAccountUrl.create, method: 'POST', data: newAccount });
   }
@@ -34,5 +46,5 @@ export function useBankAccount(): BankAccountInterface {
     return call<BankAccount>({ url: BankAccountUrl.update(id), method: 'PUT', data: changedAccount });
   }
 
-  return { getAccounts, createAccount, updateAccount };
+  return useMemo(() => ({ getAccounts, getAccount, createAccount, updateAccount }), [call]);
 }