+++ packages/react/src/hooks/fiat.hook.ts
@@ -1,11 +1,14 @@
+import { useMemo } from 'react';
 import { Fiat, FiatUrl } from '../definitions/fiat';
 import { useApi } from './api.hook';
 
 export interface FiatInterface {
   getCurrencies: () => Promise<Fiat[]>;
+  getCurrency: (currencies?: Fiat[], identifier?: string) => Fiat | undefined;
+  getDefaultCurrency: (currencies?: Fiat[]) => Fiat | undefined;
+
   toDescription: (currency: Fiat) => string;
   toSymbol: (currency: Fiat) => string;
-  getDefaultCurrency: (currencies: Fiat[]) => Fiat | undefined;
 }
 
 export function useFiat(): FiatInterface {
@@ -13,6 +16,15 @@ export function useFiat(): FiatInterface {
 
   async function getCurrencies(): Promise<Fiat[]> {
     return call<Fiat[]>({ url: FiatUrl.get, method: 'GET' });
+  }
+
+  function getCurrency(currencies: Fiat[] = [], identifier?: string): Fiat | undefined {
+    if (!identifier) return undefined;
+
+    return (
+      currencies.find((a) => a.id === +identifier) ??
+      currencies.find((a) => a.name.toLowerCase() === identifier.toLowerCase())
+    );
   }
 
   const definitions = {
@@ -30,10 +42,14 @@ export function useFiat(): FiatInterface {
     } as Record<string, string>,
   };
 
-  return {
-    getCurrencies,
-    toDescription: (currency: Fiat) => definitions.description[currency.name],
-    toSymbol: (currency: Fiat) => definitions.symbol[currency.name],
-    getDefaultCurrency: (currencies: Fiat[]) => currencies.find((f) => f.name === 'EUR'),
-  };
+  return useMemo(
+    () => ({
+      getCurrencies,
+      getCurrency,
+      getDefaultCurrency: (currencies: Fiat[] = []) => currencies.find((f) => f.name === 'EUR'),
+      toDescription: (currency: Fiat) => definitions.description[currency.name],
+      toSymbol: (currency: Fiat) => definitions.symbol[currency.name],
+    }),
+    [call],
+  );
 }